Add web UI experiments, training & utilities
Large feature addition and refactor: add experiments and training web UI pages, MLflow helpers, and many utility modules. Introduces a Streamlit export_buttons component and makes streamlit app startup more robust (patch_marking st.web / stcli handling). Improve results_visualizer to accept flexible normalize options for confusion matrices. Add CLI config/dispatch/parser shims, a source-checkout import shim (mammography/__init__.py), a CITATION.cff and README guidance for reproducibility, extensive new pipeline/io/vis/tools modules, and a large suite of tests and test outputs.

+++ src/mammography/commands/batch_inference.py
@@ -549,7 +549,7 @@ def main(argv: Sequence[str] | None = None) -> int:
                     logits = model(x)
 
                 # Convert to probabilities
-                probs = torch.softmax(logits, dim=1).cpu().numpy()
+                probs = torch.softmax(logits.float(), dim=1).cpu().numpy()
                 preds = probs.argmax(axis=1)
 
                 # Stream results for this batch
